Moved the GpuAverage change into shims
